requirements.txt
Django>=4.1.5,<4.2.0
djangorestframework>=3.14.0,<3.15.0
psycopg2>=2.9.5,<2.10.0
Pillow>=9.4.0,<9.5.0
django-debug-toolbar>=3.8.1,<3.9.0
django-cors-headers>=3.13.0,<3.14.0
djoser>=2.1.0,<2.2.0
djangorestframework_simplejwt>=4.8.0,<4.9.0
social-auth-app-django>=4.0.0,<=5.0.0
django_filter>=22.1.0,<22.2.0
celery>=5.2.7,<5.3.0
redis>=4.4.2,<4.5.0
flower>=1.2.0,<1.3.0
whitenoise>=6.3.0,<6.4.9
django-silk>=5.0.3,<5.1.0
asgiref==3.6.0
pytz==2022.7.1
#drf-nested-routers>=0.93.4,<0.99.9